+++ src/3rdparty/chromium/base/files/scoped_file_linux.cc
@@ -1,4 +1,4 @@
-// Copyright 2021 The Chromium Authors. All rights reserved.
+// Copyright 2021 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -37,7 +37,8 @@ bool CanTrack(int fd) {
 }
 
 void UpdateAndCheckFdOwnership(int fd, bool owned) {
-  if (CanTrack(fd) && g_is_fd_owned[fd].exchange(owned) == owned &&
+  if (CanTrack(fd) &&
+      g_is_fd_owned[static_cast<size_t>(fd)].exchange(owned) == owned &&
       g_is_ownership_enforced) {
     CrashOnFdOwnershipViolation();
   }
@@ -73,7 +74,7 @@ void ResetFDOwnership() {
 }  // namespace subtle
 
 bool IsFDOwned(int fd) {
-  return CanTrack(fd) && g_is_fd_owned[fd];
+  return CanTrack(fd) && g_is_fd_owned[static_cast<size_t>(fd)];
 }
 
 }  // namespace base
