+++ src/3rdparty/chromium/base/process/kill_fuchsia.cc
@@ -1,4 +1,4 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
+// Copyright 2017 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -33,7 +33,7 @@ TerminationStatus GetTerminationStatus(ProcessHandle handle, int* exit_code) {
     return TERMINATION_STATUS_STILL_RUNNING;
   }
 
-  *exit_code = process_info.return_code;
+  *exit_code = static_cast<int>(process_info.return_code);
   switch (process_info.return_code) {
     case 0:
       return TERMINATION_STATUS_NORMAL_TERMINATION;
