+++ src/3rdparty/chromium/components/dbus/menu/menu_property_list.cc
@@ -1,4 +1,4 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
+// Copyright 2019 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -55,7 +55,7 @@ std::vector<DbusString> GetDbusMenuShortcut(ui::Accelerator accelerator) {
 }  // namespace
 
 MenuItemProperties ComputeMenuPropertiesForMenuItem(ui::MenuModel* menu,
-                                                    int i) {
+                                                    size_t i) {
   // Properties should only be set if they differ from the default values.
   MenuItemProperties properties;
